Move ErrDupeLeaf to pkg directory and rename its package name.
