+++ serverless/internal/storage/fs/fs.go
@@ -27,7 +27,8 @@ import (
 	"github.com/golang/glog"
 	"github.com/google/trillian-examples/serverless/api"
 	"github.com/google/trillian-examples/serverless/api/layout"
-	"github.com/google/trillian-examples/serverless/internal/storage"
+
+	stErrors "github.com/google/trillian-examples/serverless/pkg/storage/errors"
 )
 
 const (
@@ -105,8 +106,7 @@ func Create(rootDir string) (*Storage, error) {
 // This method will attempt to silently squash duplicate leaves, but it cannot
 // be guaranteed that no duplicate entries will exist.
 // Returns the sequence number assigned to this leaf (if the leaf has already
-// been sequenced it will return the original sequence number and
-// storage.ErrDupeLeaf).
+// been sequenced it will return the original sequence number and ErrDupeLeaf).
 func (fs *Storage) Sequence(_ context.Context, leafhash []byte, leaf []byte) (uint64, error) {
 	// 1. Check for dupe leafhash
 	// 2. Write temp file
@@ -127,7 +127,7 @@ func (fs *Storage) Sequence(_ context.Context, leafhash []byte, leaf []byte) (ui
 		if err != nil {
 			return 0, err
 		}
-		return origSeq, storage.ErrDupeLeaf
+		return origSeq, stErrors.ErrDupeLeaf
 	}
 
 	// Write a temp file with the leaf data
