+++ serverless/pkg/log/integrate.go
@@ -46,7 +46,7 @@ type Storage interface {
 	//
 	// If a duplicate leaf is sequenced the storage implementation may return
 	// the sequence number associated with an earlier instance, along with a
-	// os.ErrDupeLeaf error.
+	// ErrDupeLeaf error.
 	Sequence(ctx context.Context, leafhash []byte, leaf []byte) (uint64, error)
 
 	// ScanSequenced calls f for each contiguous sequenced log entry >= begin.
